moves zonal export code to its own file
